requirements.txt
pip>=25.1.1
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.2
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
filelock==3.17.0
firebase-admin==6.6.0
fonttools==4.56.0
fsspec==2025.2.0
gcloud==0.18.3
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
kiwisolver==1.4.8
kornia==0.8.0
kornia-rs==0.1.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.29.0
networkx==3.4.2
numpy==2.2.3
oauth2client==4.1.3
opencv-python-headless
packaging==24.2
pandas==2.2.3
Pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.2.1
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
requests-toolbelt==0.10.1
rpds-py==0.23.1
rsa==4.9
setuptools==65.5.0
six==1.17.0
smmap==5.0.2
streamlit==1.43.0
sympy==1.13.1
tenacity==9.0.0
toml==0.10.2
torch==2.6.0
torchgeometry==0.1.2
torchvision==0.21.0
tornado==6.4.2
typing_extensions>=4.12.0,<4.14
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
watchdog==6.0.0
